requirements.txt
Django==2.2.3
django-rest-knox==4.1.0
djangorestframework==3.9.4
mysqlclient==1.4.2
django-cors-headers==3.0.0
gunicorn==19.9.0
whitenoise==3.3.1
